+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## Unreleased
+
+### Added
+
+* Cross-reference table reconstruction for damaged PDFs, controllable via the
+  new 'parser.try_xref_reconstruction' option
+
+
 ## 0.12.3 - 2020-08-22
 
 ### Changed